Stick with manual parsing over native (for now) (#16)
Since the native parsing APIs are really new, there's a risk of
inconsistency in their outputs. For now, use the manually-managed data
and ignore whether the native APIs exist.

This has the nice side effect of a slightly slimmer bundle, but that's
secondary to the actual goal here.

Fixes #8.

Follow-up: track the support of the native APIs and continue their use
if appropriate - #17.