Cancel autofill request when modal one starts (#29)
In some browsers (Chrome at least), trying to start a modal request for
either registration or authentication will fail if there's a pending
conditional request that's already been started. Translation:
`startAuth()` fails if `handleAutofill()` has been called. This seems to
at the very least be an ambiguity in the spec, but it doesn't happen in
Safari and Mozilla was receptive to similar feedback and [changed the
behavior in
Firefox](https://bugzilla.mozilla.org/show_bug.cgi?id=1877790).

To avoid this problem, this now adds and tracks an `AbortController` and
binds the signal to the webauthn requests. When calling any of the
methods that start a new one, anything that's pending is aborted
allowing the new one to go through.

If Chrome(ium?) changes this behavior I'd like to revert this change
since it slightly degrades the autofill behavior, but for now this seems
like the most reliable workaround that doesn't involve any client logic
or disruption.

Note that the array implementation, for reasons I don't fully
understand, seemed to be the only way to reliably a) get the old request
to cancel and b) not immediately cancel the new request. I suspect some
sort of concurrency/synchronization problem with the async nature of the
APIs. Having a simple "current controller" property did not work.