Make JsonEncodable global, and add the new getClientCapabilities API (#37)
This adds type definitions for [Client
Capabilities](https://w3c.github.io/webauthn/#sctn-getClientCapabilities),
which is (or soon will be) a replacement for the several other
promise-based APIs.

Note that other capabilities are currently in an even earlier-draft
state as of writing (still in W3 PRs or explainers) and have been
intentionally omitted. The above is a living document, so future
references may be out of date.