Add wrappers for availability APIs (#35)
Adds two new public SDK methods: `isConditionalGetAvailable():
Promise<boolean>` and `isConditionalCreateAvailable():
Promise<boolean>`. As the names suggest, they provide information about
availability of the `mediation: conditional` paths for `get()` (passkey
autofill) and `create()` (automatic passkey creation on sign in),

It is NOT necessary for clients to call these prior to using the other
SnapAuth Client APIs - but they're available to have the option of
modifying the UI based on client capabilities (as well as for internal
use).