Prepare support for automatic passkey upgrades (#36)
This adds a new API, `upgradeToPasskey()`, which is intended to be
invoked for automatic passkey upgrades after a user signs in with a
non-WebAuthn method. It's used similarly to the `handleAutofill()`,
where it's called optimistically and your callback will be invoked (or
not) if the right conditions are met.

As of today, the only supporting browser is the beta versions of Safari,
which should be publicly released in fall-ish with iOS18 and macOS
Sequoia. However, the SnapAuth API is universally available.

Fixes #39